% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grHeckmanSelect.R
\name{coef.grHeckSelect}
\alias{coef.grHeckSelect}
\title{Title Coefficients from objects returned by grHeckSelect function}
\usage{
\method{coef}{grHeckSelect}(object, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from "grHeckSelect"}

\item{...}{}
}
\value{
vector of coefficients for the selection and outcome models
}
\description{
Title Coefficients from objects returned by grHeckSelect function
}
\examples{
#coef(grHeckSelect(W=W, X=X, s=s, y=y, group_sel, group_out, penalty="grLasso"))
}
