% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grHeckmanSelect.R
\name{coef.grHeckSelect_bar}
\alias{coef.grHeckSelect_bar}
\title{Title Coefficients from objects returned by grHeckSelect_bar function}
\usage{
\method{coef}{grHeckSelect_bar}(object, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from "grHeckSelect_bar"}

\item{...}{}
}
\value{
vector of coefficients for the selection and outcome models
}
\description{
Title Coefficients from objects returned by grHeckSelect_bar function
}
\examples{
#coef(grHeckSelect_bar(W,X,s,y,group_sel, group_out, method=2))

}
