% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grHeckmanSelect.R
\name{grHeckSelect}
\alias{grHeckSelect}
\title{Group variable selection in nonignorable missing data using LASSO, SCAD and MCP.}
\usage{
grHeckSelect(
  W,
  X,
  s,
  y,
  group_sel,
  group_out,
  penalty = "grLasso",
  nlambda = 100,
  lambda,
  lambda.min = 0.001,
  log.lambda = TRUE,
  eps = 1e-04,
  max_iter = 10000,
  gamma,
  ngamma = 100,
  group_multiplier,
  penalty.factors = NULL,
  unpenalized_fixed = FALSE
)
}
\arguments{
\item{W}{A matrix of covariates in selection equation (intercept is not included).}

\item{X}{A matrix of covariates in outcome equation (intercept is not included).}

\item{s}{Binary outcome for the selection equation (0/1 and false/true allowed).}

\item{y}{Continuous outcome for the outcome equation (0 for NA).}

\item{group_sel}{A vector describing the grouping of the coefficients in the selection equation.
It is best if group is a vector of consecutive integers. If there are coefficients to be included
in the model without being penalized, assign them to group 0.}

\item{group_out}{A vector describing the grouping of the coefficients in the outcome equation.
It is best if group is a vector of consecutive integers. If there are coefficients to be included
in the model without being penalized, assign them to group 0.}

\item{penalty}{The penalty to be applied to the model, one of grLasso, grSCAD, or grMCP.}

\item{nlambda}{The number of lambda values. Default is 100.}

\item{lambda}{A user supplied sequence of lambda values.
Typically, this is left unspecified, and the function automatically
computes a grid of lambda values.}

\item{lambda.min}{The smallest value for lambda, as a fraction of lambda.max. Default is .001.}

\item{log.lambda}{When TRUE compute the grid values of lambda on log scale (default) or linear scale otherwise.}

\item{eps}{Convergence threshhold. Default is 1e-4}

\item{max_iter}{Maximum number of iterations (total across entire path). Default is 10000.}

\item{gamma}{Tuning parameter of the group MCP/SCAD penalty. Default is 3 for MCP and 4 for SCAD}

\item{group_multiplier}{A vector of values representing multiplicative factors by which each group's penalty is to be multiplied.
The default is the square root of group size.}

\item{penalty.factors}{Allows for the use of weighted versions of grLasso, grSCAD, or grMCP.}

\item{init_strat}{Use MLE to initialize the group descent algorithm otherwise use zeros.}
}
\value{
class grHeckSelect containing optimal penalized coefficients, lambda values,
bic values etc.
}
\description{
Group variable selection in nonignorable missing data using LASSO, SCAD and MCP.
}
\examples{
#grHeckSelect(W=W, X=X, s=s, y=y, group_sel, group_out, penalty="grLasso")

}
