% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grHeckmanSelect.R
\name{grHeckSelect_bar}
\alias{grHeckSelect_bar}
\title{Group variable selection in nonignorable missing data using Broken Adaptive Ridge penalty.}
\usage{
grHeckSelect_bar(
  W,
  X,
  s,
  y,
  group_sel,
  group_out,
  ridge_lambda,
  lambda,
  nridge_lambda = 10,
  nlambda = 50,
  ridge_lambda.min = 0.001,
  lambda.min = 0.001,
  log.ridge_lambda = TRUE,
  log.lambda = TRUE,
  ridge_eps = 1e-04,
  eps = 1e-04,
  inner_eps = 1e-04,
  max_iter = 10000,
  init_strat = "MLE",
  del = 0.01,
  threshold_sel = 0.005,
  threshold_out = 0.005,
  method = 2
)
}
\arguments{
\item{W}{A matrix of covariates in selection equation (intercept is not included).}

\item{X}{A matrix of covariates in outcome equation (intercept is not included).}

\item{s}{Binary outcome for the selection equation (0/1 and false/true allowed).}

\item{y}{Continuous outcome for the outcome equation (0 for NA).}

\item{group_sel}{A vector describing the grouping of the coefficients in the selection equation.
It is best if group is a vector of consecutive integers. If there are coefficients to be included
in the model without being penalized, assign them to group 0.}

\item{group_out}{A vector describing the grouping of the coefficients in the outcome equation.
It is best if group is a vector of consecutive integers. If there are coefficients to be included
in the model without being penalized, assign them to group 0.}

\item{ridge_lambda}{A user supplied sequence of lambda values for the initial
ridge estimate. Typically, this is left unspecified, and the function automatically
computes a grid of lambda values.}

\item{lambda}{A user supplied sequence of lambda values for
the BAR iteration. Typically, this is left unspecified, and the function automatically
computes a grid of lambda values.}

\item{nridge_lambda}{The number of lambda values for the ridge. Default is 10.}

\item{nlambda}{The number of lambda values for the BAR iteration. Default is 50
which leads to 500 combinations in the grid to test.}

\item{ridge_lambda.min}{The smallest value of ridge_lambda. The default is 0.001.}

\item{lambda.min}{The smallest value for lambda in BAR iteration. The default is 0.001.}

\item{log.ridge_lambda}{When TRUE compute the grid values of ridge_lambda on
log scale (default) or linear scale otherwise.}

\item{log.lambda}{When TRUE compute the grid values of lambda on
log scale (default) or linear scale otherwise.}

\item{ridge_eps}{The value fo tolerance used in intial ridge estimation.}

\item{eps}{The tolerance between BAR iteration.}

\item{inner_eps}{The value of eps for any optimization within each BAR iteration.}

\item{max_iter}{Maximum number of iterations (total across entire path). Default is 10000.}

\item{init_strat}{Use MLE to initialize the group descent algorithm otherwise use zeros.}

\item{del}{threshold_sel The value for which to set parameter values with magnitude
less than it to 0. Default is 0.005.}

\item{threshold_out}{Same as threshold_sel.}

\item{method}{Three methods are implemented. Method 2 is the fastest (see details
in accompanying paper).}
}
\value{
class grHeckSelect_bar containing optimal penalized coefficients, lambda values,
bic values etc.
}
\description{
Group variable selection in nonignorable missing data using Broken Adaptive Ridge penalty.
}
\examples{
#grHeckSelect_bar(W,X,s,y,group_sel, group_out, method=2)

}
